Profile img compress.

+++ UploadProfileImg.ashx
@@ -3,9 +3,10 @@
 using System;
 using System.Web;
 using System.IO;
+using Igprog;
 
 public class UploadProfileImg : IHttpHandler {
-
+    Global G = new Global();
     public void ProcessRequest (HttpContext context) {
         context.Response.ContentType = "text/plain";
         string userId = context.Request.Form["userid"];
@@ -15,16 +16,33 @@ public class UploadProfileImg : IHttpHandler {
             for (int i = 0; i < files.Count; i++) {
                 HttpPostedFile file = files[i];
                 string fname = context.Server.MapPath(string.Format("~/upload/users/{0}/clients/{1}/profileimg/{2}", userId, clientId, file.FileName));
+                string fname_temp = context.Server.MapPath(string.Format("~/upload/users/{0}/clients/{1}/profileimg/temp/{2}", userId, clientId, file.FileName));
                 if (!string.IsNullOrEmpty(file.FileName)) {
-                    string folderPath = context.Server.MapPath(string.Format("~/upload/users/{0}/clients/{1}/profileimg", userId, clientId));
-                    if (!Directory.Exists(folderPath)) {
-                        Directory.CreateDirectory(folderPath);
+                    int fileLength = file.ContentLength;
+                    if (fileLength <= G.KBToByte(2500)) {
+                        string folderPath = context.Server.MapPath(string.Format("~/upload/users/{0}/clients/{1}/profileimg", userId, clientId));
+                        string folderPath_temp = context.Server.MapPath(string.Format("~/upload/users/{0}/clients/{1}/profileimg/temp", userId, clientId));
+                        if (!Directory.Exists(folderPath)) {
+                            Directory.CreateDirectory(folderPath);
+                            Directory.CreateDirectory(folderPath_temp);
+                        } else {
+                            Directory.Delete(folderPath, true);
+                            Directory.CreateDirectory(folderPath);
+                            Directory.CreateDirectory(folderPath_temp);
+                        }
+                        if (fileLength <= G.KBToByte(150)) {
+                            file.SaveAs(fname);
+                        } else {
+                            file.SaveAs(fname_temp);
+                            G.CompressImage(fname_temp, file.FileName, folderPath, G.CompressionParam(fileLength));
+                        }
+                        if (Directory.Exists(folderPath_temp)) {
+                            Directory.Delete(folderPath_temp, true);
+                        }
+                        context.Response.Write(string.Format("{0}?v={1}", file.FileName, DateTime.Now.Ticks));
                     } else {
-                        Directory.Delete(folderPath, true);
-                        Directory.CreateDirectory(folderPath);
+                        context.Response.Write("max upload file size is 2.5 MB");
                     }
-                    file.SaveAs(fname);
-                    context.Response.Write(string.Format("{0}?v={1}", file.FileName, DateTime.Now.Ticks));
                 } else {
                     context.Response.Write("please choose a file to upload");
                 }
